% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/int_pairs.R
\name{int_pairs}
\alias{int_pairs}
\title{int_pairs}
\usage{
int_pairs(df)
}
\arguments{
\item{df}{(data.frame) the dataframe containing the data you want to plot}
}
\description{
Plots an interactive scattermatrix.
             You can interactively change the color of rhe points based on
             setting some threshold value and chosing which column to
             use as the one that determines color.
}
\details{
You set the threshold by using a slider from 0 - 100, representing
          percentage of the range of values that exists in the column
          selected.
}
\examples{
int_pairs(iris)
}
\author{
Ronny Restrepo
}
\seealso{
\code{\link{plot.cols}}
}
\keyword{int_pairs}
\keyword{interactive}
\keyword{manipulate}
\keyword{pairs}
\keyword{plot}

