% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/int_plot.cols.R
\name{int_plot.cols}
\alias{int_plot.cols}
\title{int_plot.cols}
\usage{
int_plot.cols(df)
}
\arguments{
\item{df}{(data.frame) the dataframe containing the data you want to plot}
}
\description{
Plots an interactive grid of scatterplots, where each cell
             plotted corresponds to data from each column of a dataframe.
             You can interactively change the color of rhe points based on
             setting some threshold value and chosing which column to
             use as the one that determines color.
}
\details{
You set the threshold by using a slider from 0 - 100, representing
          percentage of the range of values that exists in the column
          selected.
}
\examples{
int_plot.cols(iris)
}
\author{
Ronny Restrepo
}
\seealso{
\code{\link{plot.cols}}
}
\keyword{int_plot.cols}
\keyword{interactive}
\keyword{manipulate}
\keyword{plot}
\keyword{plot.cols}

