% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_convenience.R
\name{plot.cols}
\alias{plot.cols}
\title{plot.cols}
\usage{
\method{plot}{cols}(x, y = NA, type = "scatter", labelCex = 1, ...)
}
\arguments{
\item{x}{(dataframe of numerics) dataframe that will act as the x values you
want to plot}

\item{y}{(vector of numerics) OPTIONAL. vector of the outcome values.

         - If y=NA (DEFAULT) then only x values are plotted

         - If y is a vector of values, then it plots y as a function of the
           values of each column in x.}

\item{type}{(string) the type of plot to use for each cell.

     VALID VALUES WHEN y=NA :

     - "scatter" scatter plot of x values as function of row index

     -"hist" histogram of the values

     -"density"  density plot

     -"boxplot"  Box whisker plot

     VALID VALUES WHEN y IS GIVEN

     - "scatter" scatter plot

     - "lines", "|", "line", "l"  line plot}

\item{labelCex}{(numeric) Controls size of the cell labels}

\item{...}{aditional arguments to be passed on to the cell plots

         - See ?plot(), ?boxplot(), ?hist() to see what aditional arguments
           you can pass on to them.}
}
\description{
Plots a grid of subplots, where each cell plotted corresponds to
             data from each column of a dataframe.
}
\details{
You can provide just a dataframe as an argument, and it will alow you to get
         an idea of how each column of data is distributed.

         Alternatively, you can also provide a y value, and it will plot y as
         a function of each of the columns in turn.
}
\note{
Depending on whether you provide a value for y, there are different
      values for the "type" argument whcih may or may not be legal.
}
\examples{
# load some built in data
data(mtcars)
data(iris)

# Scatterplots of each column, with colors of points separated by Species Column
plot.cols(iris, col=iris$Species)

# Density plot of each column of variables
plot.cols(mtcars, type="density")

# Boxplot of each column of variables
plot.cols(iris, type="boxplot")

# Histogram each variable
plot.cols(iris, type="hist")

# output variable as a function of each predicto variable
plot.cols(mtcars[,-1], mtcars[,1])
plot.cols(iris[,-length(iris)], iris[,length(iris)], col=iris$Species)
}
\author{
Ronny Restrepo
}

