% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_by_cat.R
\name{plot_by_cat}
\alias{plot_by_cat}
\title{plot_by_cat}
\usage{
plot_by_cat(x, y, cat, col = "blue", alpha = 0.3, overlap = FALSE,
  hlims = TRUE)
}
\arguments{
\item{x}{(vector of numerics) The independent variable}

\item{y}{(vector of numerics) The dependent variable}

\item{cat}{(factors) The factors to do grouping with}

\item{col}{Colors of the scatter points}

\item{overlap}{(logical) Overlap the subplots over each other to create a
       single plot that is color coded for each category?

       If FALSE, then each category is plotted as a separate subplot.}

\item{hlims}{(logical) Homogenous Limits?

     TRUE = each subplot is set up with the exact same x and y limits, to
            make it easy to compare each one

     FALSE = each subplot automatically scales its limits to fill up the
             space from its own data points

     (DEFAULT is TRUE)}
}
\description{
Draws a visual separation for how y relates to x when grouped
             by some category variable. You can separate by color coding
             into the same plot (overlaping) or by separating into several
             subplots.
}
\details{
You can specify if the subplots all share the same x and y limits
         (default), or if you want each subplot to automatically scale to
         fit its sub-data with the least amount of shite space.
}
\examples{
data(mtcars)
col = c("magenta", "blue", "green")

# Separate subplot for each category (number of cylinders)
plot_by_cat(mtcars$disp, mtcars$mpg, cat=mtcars$cyl, overlap=F, col=col)

# Scatter points overlapped, but color coded for different cylinders
plot_by_cat(mtcars$disp, mtcars$mpg, cat=mtcars$cyl, overlap=T, col=col)
}
\author{
Ronny Restrepo
}
\seealso{
plot.cols plot
}
\keyword{category}
\keyword{group}
\keyword{plot_by_cat}

