% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_distribution}
\alias{plot_distribution}
\title{plot_distribution}
\usage{
plot_distribution(dist = "normal", mean = NA, sd = NA, n = NA, p = NA,
  df = NA, df2 = NA, rate = NA, res = 100, return.df = FALSE,
  primary = TRUE, p.lower = 1e-04, p.upper = 0.9999, ...)
}
\arguments{
\item{dist}{"normal" "poisson" "binomial" "t" "f" "exp" determines the
distribution to use.}

\item{mean}{numeric. The mean of the distribution. If using poisson, this
is ths lambda value.
DEFAULT = 0 if using normal distribution.
DEFAULT = 1 if using poisson distribution.}

\item{sd}{Standard deviation.
DEFAULT = 1}

\item{n}{int. Used when dealing with binomial distribution. The number
of times we run the binomial event, eg flip a coin.
DEFAULT: 1}

\item{p}{numeric. probability of success when using binomial
deistribution
DEFAULT: 0.5}

\item{df}{Degrees of Freedom when using t distribution.
Or the first degrees of freedom when using f distribution
DEFAULT = 1   (if "t" distribution chosen)
DEFAULT = 10  (if "f" distribution chosen)}

\item{df2}{Second degrees of freedom when using f distribution
DEFAULT = 100}

\item{rate}{Used for exponential Distribution
DEFAULT = 1}

\item{res}{integer. Resolution of the plot (measured as the number of
data points along the x axis)
Not implemented for Poisson distribution yet.
DEFAULT = 100 if using normal distribution}

\item{return.df}{should it return a dataframe of the x and y values?
DEFAULT = FALSE}

\item{primary}{boolean. Whether to plot as primary plot using plot() or
append to an exisitng plot using points()
DEFAULT = TRUE}

\item{p.lower}{numeric. A quantile used to calculate the lower end of the
x axis to plot. eg, if a value of 0.25 is used, then the x
axis will start at the 25th percentile of the distribution.
DEFAULT = 0.0001}

\item{p.upper}{numeric. quantile used to calculate the upper end of the
x axis to plot.
DEFAULT = 0.9999}

\item{...}{other parameters to pass onto the plot}

\item{show.mean}{Show a vertical line highlihgting mean of the distribution?
DEFAULT = TRUE}
}
\description{
Plots what a particular distribution looks like given parameters such as
- mean and standard deviation (for normal distribution).
- size and p (for binomial distribution)
- df (for t distribution)
- df and df2 (for f distribution)
- mean (for poisson distribution)
}
\examples{
plot_distribution("normal", mean=50, sd=15)
plot_distribution("binomial", n=15, p=0.9)
plot_distribution("t", df=12)
plot_distribution("exp", rate=0.2)
plot_distribution("f", df=10, df2=100)
plot_distribution("poisson", mean=50)
}
\author{
Ronny Restrepo
}

