% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_models.R
\name{plot_models}
\alias{plot_models}
\title{plot_models}
\usage{
plot_models(models, scatter = TRUE, scatter_color = "blue",
  scatter_alpha = 0.2, scatter_size = 2, model_color = "darkorange",
  model_alpha = 1, modelWidth = 2, ci = FALSE, ci_level = 0.95,
  ci_color = model_color, ci_alpha = model_alpha/4, draw_grid = TRUE, ...)
}
\arguments{
\item{models}{a list of model objects}

\item{scatter}{(logical)

     TRUE = plots scatterplot of the datapoints used in the training set.

     FALSE =  only plots the models without any of the original data points.}

\item{scatter_color}{the color for the scatter plot of the original data.

     See here for a chart of possible colors:
     \url{http://research.stowers-institute.org/efg/R/Color/Chart/}}

\item{scatter_alpha}{(numeric) The alpha of the scatter plots}

\item{scatter_size}{(numeric) Size of the scatter points (Default is 2)}

\item{model_color}{either a single color for all models to be. Or a vector
of colors to cycle through.}

\item{model_alpha}{(numeric) The alpha of the model lines}

\item{modelWidth}{(numeric) width of the model lines. You can wither
pass a single value in which case all models will be the same width,
or you can specify a vector of numerics to  specify the width of each
individual model.}

\item{ci}{(logical) Plot the Confidence Interval?

     (Default is TRUE)}

\item{ci_level}{(numeric) confidence level to use. Must be a value between
     0 and 1.

     (Default is 0.95)}

\item{ci_color}{Color of the Confidence Interval Shaded Region

     (Default is to use same color as model_color)}

\item{ci_alpha}{(numeric) Alpha level of Confidence Interval Shaded Region.

     (Default is model_alpha/4)}

\item{draw_grid}{(logical) should it include a grid?

     (Default is TRUE)}

\item{...}{other arguments to be passed on to the parent plot() function.

         use ?plot() to see what other arguments can be passed on.}
}
\description{
Plots multiple models fitted to the same data. By default it
         overlaps each model over the others. But you can specify to have
         each in a separate cell in a grid plot.
}
\details{
By default the plot of each model will be the same color. But you
         can specify the colour of each individualy.

         Alternatively, a feature that will be added is to have all plots
         the same color except for a master model (needs to be placed as the
         first model in the list of models)
         , or to have a master and a set of slaves (where
         one is brightly colored, and the others are grey)

         You can make the models plot with a shaded Confidence Interval. By
         default, the confidence interval is 0.95, but you can set it to
         something else.
}
\note{
Make sure that when you are creating models, that you avoid using the
      "$" operator.

      BAD:  lm(myData$y ~ myData$x)

      GOOD: lm(y ~ x, data=myData)

      Otherwise you will get error messages like  "'x' and 'y' lengths differ"
      when you try running the predict() function which is used by
      plot_models()
}
\examples{
library(ElemStatLearn)
data(ozone,package="ElemStatLearn")

# create models
model1 = loess(temperature ~ ozone, data=ozone,span=0.15)
model2 = lm(temperature ~ ozone, data=ozone)
model3 = loess(temperature ~ ozone, data=ozone, span=0.2)
model4 = loess(temperature ~ ozone, data=ozone, span=1)
model5 = loess(temperature ~ ozone, data=ozone, span=20)
models = list(model1, model2, model3, model4, model5)

plot_models(models)
plot_models(models, xlab="Ozone", ylab="Temperature",
            main="Models of Temperature as function of Ozone")

# Use different colors for each model
modcols = c("darkorchid2","chartreuse","brown2","aquamarine2","darkorange")

# Use different line widths for each model
plot_models(models, model_color=modcols, modelWidth=c(8,2,4,8,3))

# Plot models with Confidence intervals
models = list(model1, model2)
modcols = c("darkorchid2", "darkorange")
plot_models(models, model_color=modcols, ci=T)
}
\author{
Ronny Restrepo
}
\seealso{
\code{\link{plot_residuals}} \code{\link{plot.cols}} \code{\link{plot}}
}
\keyword{model}
\keyword{modelling}
\keyword{models}
\keyword{plot}
\keyword{plot.convenience}
\keyword{plot_models}
\keyword{plotting}

