% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_residuals.R
\name{plot_residuals}
\alias{plot_residuals}
\title{plot_residuals}
\usage{
plot_residuals(model, points = TRUE, lines = FALSE, ctm = FALSE,
  point_col = "blue", point_alpha = 0.3, point_size = 2,
  line_col = point_col, line_alpha = 0.3, line_width = 10 * point_size,
  ctm_point_col = "coral", ctm_point_alpha = point_alpha,
  ctm_point_size = 2, ctm_line_col = ctm_point_col,
  ctm_line_alpha = line_alpha, ctm_line_width = 10 * ctm_point_size,
  draw_grid = TRUE, ...)
}
\arguments{
\item{model}{(model object) the model you want to draw the residuals for}

\item{points}{(logical) Draw the residuals as points?

     (Default is TRUE)}

\item{lines}{(logical) Draw the residuals as lines?

     (Default is FALSE)}

\item{ctm}{(logical) Compare To the Mean? Compare the residuals left over
     from the model compared to the residuals that would exist if the mean
     of the putcome variable was used as the estimator.

     (Default is FALSE)}

\item{point_col}{Color of the points used for residuals from the model.

     See here for a chart of possible colors:
     \url{http://research.stowers-institute.org/efg/R/Color/Chart/}

     (Default is "blue")}

\item{point_alpha}{(numeric) Alpha of points used for residuals from model.

     (Default is 0.3)}

\item{point_size}{(numeric) Size of points used for residuals from the model.

     (Default is 2)}

\item{line_col}{Color of the lines used for residuals from the model.

     (Default is same as point_col)}

\item{line_alpha}{(numeric) Alpha of lines used for residuals from the model.

     (Default is 0.3)}

\item{line_width}{(numeric) Width of lines used for residuals from the model.

     (Default is a value that gives same visual thickness as point_size)}

\item{ctm_point_col}{Color of the points used for residuals from the mean.

     (Default is "coral")}

\item{ctm_point_alpha}{Alpha of the points used for residuals from the mean.

     (Default is the same as point_alpha)}

\item{ctm_point_size}{(numeric) Size of the points used for residuals from
     the mean.

     (Default is 2)}

\item{ctm_line_col}{Color of the lines used for residuals from the mean.

     (Default is same value as ctm_point_col)}

\item{ctm_line_alpha}{Alpha of the lines used for residuals from the mean.

     (Default is same value as line_alpha)}

\item{ctm_line_width}{Width of the lines used for residuals from the mean.

     (Default is a value that gives same visual thickness as ctm_point_size)}

\item{draw_grid}{(logical) Should a grid be drawn?

     (Default is TRUE)}

\item{...}{other arguments to be passed on to the parent plot() function.

         use ?plot() to see what other arguments can be passed on.}
}
\description{
Create a plot of observed x values vs residuals for some
             particular model object.
}
\details{
You can chose to either display the residuals as points, or as
         vertical line segments, or both by setting the arguments "points",
         and "lines" to TRUE or FALSE.
}
\note{
Make sure that when you are creating models, that you avoid using the
      "$" operator.

      BAD:  lm(myData$y ~ myData$x)

      GOOD: lm(y ~ x, data=myData)

      Otherwise you will get error messages like  "'x' and 'y' lengths differ"
      when you try running the predict() function which is used by
      plot_models()
}
\examples{
library(ElemStatLearn)
data(ozone,package="ElemStatLearn")

# create model
model = loess(temperature ~ ozone, data=ozone,span=0.15)

# Plot the residuals as points
plot_residuals(model)

# Plot the residuals as lines only
plot_residuals(model2, points=F, lines=T, line_width=10)

# Plot the residuals as points and lines
plot_residuals(model, lines=T)

# Plot the residuals of the model, compared to the residuals when the mean
# is used as the predictor.
plot_residuals(model, ctm=T, lines=T, points=T, point_size=1)
}
\author{
Ronny Restrepo
}
\seealso{
\code{\link{plot_models}} \code{\link{plot.cols}} \code{\link{plot}}
}
\keyword{plot}
\keyword{plot.convenience}
\keyword{plot_residuals}
\keyword{residuals}

