% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_surrounding_residuals.R
\name{plot_surrounding_residuals}
\alias{plot_surrounding_residuals}
\title{plot_surrounding_residuals}
\usage{
plot_surrounding_residuals(x, center = mean(x), dev = 2 * sd(x),
  type = "SSE", res = 100, ...)
}
\arguments{
\item{x}{(vector of numerics) The empirical values}

\item{center}{(numeric) The value you wish to center you search around}

\item{dev}{(numeric) How far to the left and right should the search span}

\item{type}{(string) What measure of error should it plot?

       "SSE"    # Sum of Squared Errors

       "MSE"    # Mean Squared Error

       (DEFAULT = "SSE")}

\item{res}{(integer)  Sets the resolution of the plot. Calculated as the
     number of points that lie between (center - dev) and  (center + dev).

     DEFAULT = 100}

\item{...}{other arguments to be passed to plot}
}
\description{
Creates a plot of what the residuals are for values surrounding
     a central point that you specify. You specify how much to deviate from
     the central point, and what measure of error you want.
}
\details{
This can be used to visualise whether some value you picked out is
indeed a local minimum, or it can be used to visually explore possible values
via an informed trial and error (where the plot may guide you in which
direction to follow). You chose how wide to make your search through the
\code{"dev"} argument.
}
\examples{
x <- c(12,11,13,15,16,12,11,9,14,11,10,7,15)
plot_surrounding_residuals(x)
plot_surrounding_residuals(x, center=7, dev=10)
plot_surrounding_residuals(x, center=13, dev=2, type="MSE")
plot_surrounding_residuals(x, center=13, dev=2, type="SSE")
}
\author{
Ronny Restrepo
}

