% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_par_for_n_subplots.R
\name{set_par_for_n_subplots}
\alias{set_par_for_n_subplots}
\title{Helper function to set global parameters of par() that are appropriate for
plotting n number of cells.}
\usage{
set_par_for_n_subplots(n)
}
\description{
You secify n, and it automatically calculates what size the grid, and other
plotting parameters to generate a nice layout.
}

