% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_convenience.R
\name{shade_between}
\alias{shade_between}
\title{shade_between}
\usage{
shade_between(x, y, lower, upper, primary = TRUE, shade.col = "red",
  shade.border = NULL, shade.lty = par("lty"), shade.density = NULL,
  shade.angle = 45, ...)
}
\arguments{
\item{x}{(vector) x coordinates of the plot}

\item{y}{(vector) y coordinates of the plot}

\item{lower}{(numeric) lower point (Output snaps to closest element that is
actually in x)}

\item{upper}{(numeric) uppper point (Output snaps to closest element that is
actually in x)}

\item{primary}{(logical) Should it plot as a new plot (TRUE), or
         overlay on top of an existing plot (FALSE)

         (DEFAULT = TRUE)}

\item{shade.col}{(string) color to fill in the shaded area.}

\item{shade.border}{(string, NA, NULL, logical) the color to draw the border.
         The default,  NULL, means to use par("fg").

         Use border = NA to omit borders.

         For compatibility with S, border can also be logical, in
         which case FALSE is equivalent to NA (borders omitted)

         TRUE is equivalent to NULL (use foreground colour)}

\item{shade.lty}{(string) the line type to be used for the shaded area}

\item{shade.density}{(numeric) the density of shading lines, in lines per
         inch. The  default value of NULL means that no shading lines are
         drawn. A zero value of density means no shading nor
         filling whereas negative values and NA suppress shading
         (and so allow color filling).}

\item{shade.angle}{(integer) the slope of shading lines, given as an angle
in degrees (counter-clockwise).}

\item{...}{aditional parameters for the plot, such as type, lty, etc}
}
\description{
Creates a plot which shades the area under the curve between "lower" and
"upper" points along x.
}
\details{
Note, that depending on the resolution of the x points, you may not
actually get the shaded region in the exact upper and lower points you
specified. The edges of the shaded region will snap to the closest value
that is actually in the x vector.
}
\examples{
# Shades between -5 and -2 in a bell curve
   x = seq(-9,9,by=0.01)
   y = dnorm(x, mean=0, sd=3)
   shade_between(x,y,-5,-2, type="l", lwd=3, shade.density=20)
}

