% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.cols.R
\name{plot.cols}
\alias{plot.cols}
\title{plot.cols}
\usage{
\method{plot}{cols}(x, y = NA, type = "scatter", labelCex = 1,
  col = "darkgray", grad = FALSE, grad.theme = "flame",
  grad.scal = "normal", ...)
}
\arguments{
\item{x}{(dataframe of numerics) dataframe that will act as the x values you
want to plot}

\item{y}{(vector of numerics) OPTIONAL. vector of the outcome values.

         - If \code{y=NA} (DEFAULT) then only \code{x} values are plotted

         - If \code{y} is a vector of values, then it plots y as a function of the
           values of each column in \code{x}.}

\item{type}{(string) the type of plot to use for each cell.

     VALID VALUES WHEN \code{y=NA} :

     - \code{"scatter"} scatter plot of x values as function of row index

     -\code{"hist"} histogram of the values

     -\code{"density"}  density plot

     -\code{"boxplot"}  Box whisker plot


     VALID VALUES WHEN \code{y} IS GIVEN

     - \code{"scatter"} scatter plot

     - \code{"lines"}, \code{"|"}, \code{"line"}, \code{"l"}  line plot}

\item{labelCex}{(numeric) Controls size of the cell labels}

\item{col}{the color(s) of the points/lines of the plot. You can use whatever
       value you would pass on to \code{plot()}

       You can pass an individual value, or a vector of values, eg you can
       specify that the color of the points be determined by the value of
       the output variable in your data.}

\item{grad}{(logical)  Should gradient colours be used?

       \code{TRUE} - If you specified a vector of numeric values for col,
              then you can chose to have the color be a gradient instead of
              discrete colors.
       \code{FALSE} - (DEFAULT) values in 'col' will be interpreted as
              discrete color changes.}

\item{grad.theme}{(string) Controls the gradient color

       \code{"flame"} = from yellow to red

       \code{"blue"} = from light blue to dark blue

       \code{"rainbow"} = From blue,cyan, green, yellow, orange, red

        anything else = from light gray to black.}

\item{grad.scal}{(string) EXPERIMENTAL - controls how the gradient is
        inerpolated.

        \code{"normal"} normally distributed

        \code{"range"}  scales it linearly from minimum to maximum value

        Please note that this argument is experimental and may be removed
        at any point.}

\item{...}{aditional arguments to be passed on to the cell plots

         - See \code{?plot()}, \code{?boxplot()}, \code{?hist()} to see what
           aditional arguments you can pass on to them.}
}
\description{
Plots a grid of subplots, where each cell plotted corresponds to
             data from each column of a dataframe.
}
\details{
You can provide just a dataframe as an argument, and it will alow
         you to get an idea of how each column of data is distributed.

         Alternatively, you can also provide a y value, and it will plot y as
         a function of each of the columns in turn.
}
\note{
Depending on whether you provide a value for \code{y}, there are
      different values for the \code{type} argument whcih may or may not be
      legal.
}
\examples{
# load some built in data
data(mtcars)
data(iris)

# Scatterplots of each column, with colors separated by Species Column
plot.cols(iris, col=iris$Species, pch=19)

# Scatterplots of each column, with color gradient based on mpg Column
plot.cols(mtcars, col=mtcars$mpg, grad=T, pch=19)

# Density plot of each column of variables
plot.cols(mtcars, type="density")

# Boxplot of each column of variables
plot.cols(iris, type="boxplot")

# Histogram each variable
plot.cols(iris, type="hist")

# output variable as a function of each predicto variable
plot.cols(mtcars[,-1], mtcars[,1])
plot.cols(iris[,-length(iris)], iris[,length(iris)], col=iris$Species)
}
\author{
Ronny Restrepo
}
\seealso{
\code{\link{plot}}, \code{\link{boxplot}}, \code{\link{hist}}
}
\keyword{column,}
\keyword{columns}
\keyword{plot,}
\keyword{plot.cols,}
\keyword{plotting,}

