% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_residuals.R
\name{plot_residuals}
\alias{plot_residuals}
\title{plot_residuals}
\usage{
plot_residuals(model, points = TRUE, lines = FALSE, point_col = "blue",
  point_alpha = 0.3, point_size = 2, line_col = "orange",
  line_alpha = 0.3, line_width = 5, draw_grid = TRUE, ...)
}
\arguments{
\item{model}{(model object) the model you want to draw the residuals for}

\item{points}{(logical) Draw the residuals as points?

     (Default is TRUE)}

\item{lines}{(logical) Draw the residuals as lines?

     (Default is FALSE)}

\item{point_col}{the color of the points

     (Default is "Blue")}

\item{point_alpha}{(numeric) the alpha of the points.

     (Default is 0.3)}

\item{point_size}{(numeric) The size of the points.

     (Default is 2)}

\item{line_col}{The color od the lines.

     (Default is "orange")}

\item{line_alpha}{(numeric) the alpha of the lines.

     (Default is 0.3)}

\item{line_width}{(numeric) The width of the lines.

     (Default is 5)}

\item{draw_grid}{(logical) Should a grid be drawn?

     (Default is TRUE)}

\item{...}{other arguments to be passed on to the parent plot() function.

         use ?plot() to see what other arguments can be passed on.}
}
\description{
Create a plot of observed x values vs residuals for some
             particular model object.
}
\details{
You can chose to either display the residuals as points, or as
         vertical line segments, or both by setting the arguments "points",
         and "lines" to TRUE or FALSE.
}
\note{
Make sure that when you are creating models, that you avoid using the
      "$" operator.

      BAD:  lm(myData$y ~ myData$x)

      GOOD: lm(y ~ x, data=myData)

      Otherwise you will get error messages like  "'x' and 'y' lengths differ"
      when you try running the predict() function which is used by
      plot_models()
}
\examples{
library(ElemStatLearn)
data(ozone,package="ElemStatLearn")

# create model
model = loess(temperature ~ ozone, data=ozone,span=0.15)

# Plot the residuals as points
plot_residuals(model)

# Plot the residuals as lines
plot_residuals(model2, points=F, lines=T, line_width=10)

# Plot the residuals as lines and Points
plot_residuals(model3, points=T, lines=T)
}
\author{
Ronny Restrepo
}
\seealso{
\code{\link{plot_models}} \code{\link{plot.cols}} \code{\link{plot}}
}
\keyword{plot}
\keyword{plot.convenience}
\keyword{plot_residuals}
\keyword{residuals}

